requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
albucore==0.0.21
albumentations==1.4.23
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
av==13.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
eval_type_backport==0.2.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
huggingface-hub==0.26.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
moviepy==2.0.0
mpmath==1.3.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.2
pipreqs==0.5.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rootutils==1.0.7
rpds-py==0.21.0
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools==75.1.0
simsimd==6.2.1
six==1.16.0
smmap==5.0.1
some-package==0.1
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.11.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tinycss2==1.4.0
tokenizers==0.20.3
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchmetrics==1.6.0
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.35
ultralytics-thop==2.0.12
urllib3==2.2.3
wandb==0.18.7
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.44.0
yarg==0.1.9
yarl==1.18.0
